apps/engine/requirements.txt
google-api-python-client
google-auth-oauthlib
beautifulsoup4
python-dotenv
supabase
pytest
pytest-asyncio
instructor
openai
anthropic
pydantic
google-generativeai
